desktop-agent/requirements.txt
keyboard==0.13.5
mss==9.0.1
numpy==1.26.4
openai>=1.63.0
python-dotenv==1.0.1
requests==2.32.3
httpx<0.28
sounddevice==0.4.7
faster-whisper==1.1.0
pyannote.audio==3.1.1
torch==2.3.1
torchaudio==2.3.1
pyinstaller==6.6.0
